Create Project:Tip_Calculator.py
Day 2 Project: Tip Calculator
Introduction to the Tip Calculator Project
Welcome to the final project of today. In this project, we are going to build a Tip Calculator. The final output will look like this:

It will print, "Welcome to the tip calculator." Then it will ask for the total bill amount.

For example, if you and a couple of friends went out for lunch and the bill came to $124.56, it will then ask, "What percentage tip would you like to give?"

If you want to give a 12% tip, you enter 12 and hit Enter. Finally, it will ask, "How many people are splitting this bill?"

Let's say there are seven people; after entering this, the program will tell us that each person should pay about $19.93.

Notice that even though the calculations might produce numbers with many decimal places, the final amount is rounded to two decimal places for accuracy.